Notifications for work now get through
For some reason, the startum client doesn't actually give any data about
the jobs that the server sends back, which are the only part that we need
to actually mine anything. However, the information can be taken from the
raw socket.

Using the socket and listening for data we recieve, we then can emit a
custom event and insert our own hacked logic to get notifications sent out
in a way that works for our program. So now we can keep track of several
jobs, and we just need a miner to actually work on those jobs.