Added fixes and moved main
Main has now been moved to the root directory. Additionally, we can now
pass in a few options that allow us to specify our wallet, and eventually
some other options.

For package.json, we now have the main executable specified, and so we can
run it anywhere if we install it globally.